requirements.txt
asgiref==3.9.1
Django==5.2.6
django-rest==0.8.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
